+++ src/main/java/org/octri/omop_annotator/repository/omop/VisitOccurrenceRepository.java
@@ -28,7 +28,7 @@ public interface VisitOccurrenceRepository extends PagingAndSortingRepository<Vi
 			+ " where po.person.id = ?1"
 			+ " and lower(procedure.name) like ?2"
 			+ " order by po.visitOccurrence.id asc")
-	List<Long> findByPersonIdAndProcedureNameLike(Long personId, String procedureName);
+	List<Integer> findByPersonIdAndProcedureNameLike(Integer personId, String procedureName);
 
 	@Query(value = "select distinct co.visitOccurrence.id"
 			+ " from ConditionOccurrence co"
@@ -36,7 +36,7 @@ public interface VisitOccurrenceRepository extends PagingAndSortingRepository<Vi
 			+ " where co.person.id = ?1"
 			+ " and lower(condition.name) like ?2"
 			+ " order by co.visitOccurrence.id asc")
-	List<Long> findByPersonIdAndConditionNameLike(Long personId, String conditionName);
+	List<Integer> findByPersonIdAndConditionNameLike(Integer personId, String conditionName);
 
 	@Query(value = "select distinct obs.visitOccurrence.id"
 			+ " from Observation obs"
@@ -44,7 +44,7 @@ public interface VisitOccurrenceRepository extends PagingAndSortingRepository<Vi
 			+ " where obs.person.id = ?1"
 			+ " and lower(observation.name) like ?2"
 			+ " order by obs.visitOccurrence.id asc")
-	List<Long> findByPersonIdAndObservationNameLike(Long personId, String conditionName);
+	List<Integer> findByPersonIdAndObservationNameLike(Integer personId, String conditionName);
 
 	@Query(value = "select distinct m.visitOccurrence.id"
 			+ " from Measurement m"
@@ -52,12 +52,12 @@ public interface VisitOccurrenceRepository extends PagingAndSortingRepository<Vi
 			+ " where m.person.id = ?1"
 			+ " and lower(measurement.name) like ?2"
 			+ " order by m.visitOccurrence.id asc")
-	List<Long> findByPersonIdAndMeasurementNameLike(Long personId, String conditionName);
+	List<Integer> findByPersonIdAndMeasurementNameLike(Integer personId, String conditionName);
 
 	@Query(value = "select distinct note.visitOccurrence.id"
 			+ " from Note note"
 			+ " where note.person.id = ?1"
 			+ " and lower(note.text) like ?2"
 			+ " order by note.visitOccurrence.id asc")
-	List<Long> findByPersonIdAndNoteTextLike(Long personId, String conditionName);
+	List<Integer> findByPersonIdAndNoteTextLike(Integer personId, String conditionName);
 }
